Don't show XCTest failures under Problems view
Stop parsing diagnostics once build is complete

Issue: #925
